ref(profiling): prepare for new public continuous API (#3833)
